requirements.txt
discord.py~=2.5.0
emoji~=2.14.0
pydantic~=2.11.0
pymongo~=4.13
requests~=2.32.3
types-emoji~=2.1.0
